+++ src/EngineFactory.php
@@ -6,6 +6,7 @@ namespace hanneskod\readmetester;
 
 use hanneskod\readmetester\Example\ExampleFactory;
 use hanneskod\readmetester\Example\FilterInterface;
+use hanneskod\readmetester\Example\NullFilter;
 use hanneskod\readmetester\Parser\Parser;
 use hanneskod\readmetester\Expectation\ExpectationEvaluator;
 use hanneskod\readmetester\Expectation\ExpectationFactory;
@@ -13,13 +14,14 @@ use hanneskod\readmetester\Runner\EvalRunner;
 
 class EngineFactory
 {
-    public function createEngine(FilterInterface $filter = null): Engine
+    public function createEngine(FilterInterface $filter = null, bool $ignoreUnknownAnnotations = false): Engine
     {
         return new Engine(
             new Parser,
             new ExampleFactory(
                 new ExpectationFactory,
-                $filter
+                $filter ?: new NullFilter,
+                $ignoreUnknownAnnotations
             ),
             new ExampleTester(
                 new EvalRunner,